Rename VotingCog to Voting, Tie votings to server
Renamed the VotingCog to Voting to provide a nicer help message
with the default help command.
Furthermore, votings are now tied to the server they were created on
in order to prevent uid conflicts. Evaluated votes are now automatically
deleted after 5 minutes to allow for the uid to be reused.